+++ src/gui/startup/StartupInterface.linux.cpp
@@ -26,42 +26,39 @@
  * version.  If you delete this exception statement from all source
  * files in the program, then also delete it here.
  */
-#include "StartupInterface.h"
 #include <QFileInfo>
 
+#include "StartupInterface.h"
+
 struct LinuxStartupInterface {
-	QString desktop_file_path;
+  QString desktop_file_path;
 };
 
-StartupInterface::StartupInterface(QObject* parent) :
-		QObject(parent) {
-	interface_impl_ = new LinuxStartupInterface();
+StartupInterface::StartupInterface(QObject* parent) : QObject(parent) {
+  interface_impl_ = new LinuxStartupInterface();
 
-	if(qEnvironmentVariableIsSet("HOME")) {
-		static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path = qgetenv("HOME");
-		static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path.append(QStringLiteral("/.config/autostart/Librevault.desktop"));
-	}
+  if (qEnvironmentVariableIsSet("HOME")) {
+    static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path = qgetenv("HOME");
+    static_cast<LinuxStartupInterface*>(interface_impl_)
+        ->desktop_file_path.append(QStringLiteral("/.config/autostart/Librevault.desktop"));
+  }
 }
 
-StartupInterface::~StartupInterface() {
-	delete static_cast<LinuxStartupInterface*>(interface_impl_);
-}
+StartupInterface::~StartupInterface() { delete static_cast<LinuxStartupInterface*>(interface_impl_); }
 
-bool StartupInterface::isSupported() {
-	return true;
-}
+bool StartupInterface::isSupported() { return true; }
 
 bool StartupInterface::isEnabled() const {
-	QFileInfo desktop_file(static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path);
-	return desktop_file.exists() && desktop_file.isFile();
+  QFileInfo desktop_file(static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path);
+  return desktop_file.exists() && desktop_file.isFile();
 }
 
 void StartupInterface::enable() {
-	if(!isEnabled())
-		QFile::copy(QStringLiteral(":/platform/Librevault.desktop"), static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path);
+  if (!isEnabled())
+    QFile::copy(QStringLiteral(":/platform/Librevault.desktop"),
+                static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path);
 }
 
 void StartupInterface::disable() {
-	if(isEnabled())
-		QFile::remove(static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path);
+  if (isEnabled()) QFile::remove(static_cast<LinuxStartupInterface*>(interface_impl_)->desktop_file_path);
 }
